Fixed memory errors

+++ aicontrollers/gentlemanbot.cpp
@@ -115,7 +115,7 @@ auto GentlemanBot::getDiscard() -> Mahjong::Piece{
 
   for(const auto &p : preferedDiscards){
     if(p.isHonor()){
-      for(int i=0; i<hand.size(); i++){
+      for(size_t i=0; i<hand.size(); i++){
         if(hand[i].toUint8_t() == p.toUint8_t()){
           hand.erase(hand.begin()+i);
         }
@@ -123,7 +123,7 @@ auto GentlemanBot::getDiscard() -> Mahjong::Piece{
       return p;
     }
   }
-  for(const auto &p : preferedDiscards){for(int i=0; i<hand.size(); i++){
+  for(const auto &p : preferedDiscards){for(size_t i=0; i<hand.size(); i++){
         if(hand[i].toUint8_t() == p.toUint8_t()){
           hand.erase(hand.begin()+i);
         }
@@ -131,7 +131,7 @@ auto GentlemanBot::getDiscard() -> Mahjong::Piece{
       return p;
   }
   for(const auto &p : secondTierDiscards){
-    for(int i=0; i<hand.size(); i++){
+    for(size_t i=0; i<hand.size(); i++){
         if(hand[i].toUint8_t() == p.toUint8_t()){
           hand.erase(hand.begin()+i);
         }
@@ -139,7 +139,7 @@ auto GentlemanBot::getDiscard() -> Mahjong::Piece{
       return p;
   }
   for(const auto &p : thirdTierDiscards){
-    for(int i=0; i<hand.size(); i++){
+    for(size_t i=0; i<hand.size(); i++){
         if(hand[i].toUint8_t() == p.toUint8_t()){
           hand.erase(hand.begin()+i);
         }
